gcp/viet-soccer/firestore-api/requirements.txt
Flask==2.0.0
gunicorn==20.1.0
grpcio==1.37.1
protobuf==3.17.0
google-auth==1.30.1
google-cloud-firestore==2.1.1
googleapis-common-protos==1.53.0
